major: rename to iterpy to align with pypi
