+++ iterpy/_generate_pyi.py
@@ -21,8 +21,8 @@ if __name__ == "__main__":
 
     combined_interface = f"    # Code for generating the following is in {Path(__file__).name}"
     for mark in [
-        TypeMarker("Iterable[S]"),
-        TypeMarker("Iterator[S]"),
+        TypeMarker("iterpyable[S]"),
+        TypeMarker("iterpyator[S]"),
         TypeMarker("tuple[S, ...]"),
         TypeMarker("Sequence[S]"),
         TypeMarker("list[S]"),
