+++ iterpy/_iter.pyi
@@ -4,18 +4,18 @@ This type stub file was generated by pyright.
 
 from collections.abc import (
     Callable,
-    Iterable,
-    Iterator,
+    iterpyable,
+    iterpyator,
     Sequence,
 )
-from typing import Generic, TypeVar, overload
+from typing import Generic, Iterable, Iterator, TypeVar, overload
 
 S = TypeVar("S")
 T = TypeVar("T")
 U = TypeVar("U")
 
-Iterables: TypeVar = (
-    Iterable | Iterator | tuple | list | set | frozenset  # type: ignore
+iterpyables: TypeVar = (
+    iterpyable | iterpyator | tuple | list | set | frozenset  # type: ignore
 )
 
 class Iter(Generic[T]):
@@ -27,7 +27,7 @@ class Iter(Generic[T]):
     def count(self) -> int: ...
     def to_list(self) -> list[T]: ...
     def to_tuple(self) -> tuple[T, ...]: ...
-    def to_iter(self) -> Iterator[T]: ...
+    def to_iterator(self) -> Iterator[T]: ...
     def to_set(self) -> set[T]: ...
     def map(self, func: Callable[[T], S]) -> Iter[S]:  # noqa: A003
         ...
