+++ iterpy/benchmark/query_1/polars_q1.py
